refactor ForgotPassword to separate data source for avoid problems with Segregation Interface principle with IAuthenticationDataSource
