+++ app/src/main/java/org/eyeseetea/malariacare/data/authentication/AuthenticationManager.java
@@ -8,6 +8,7 @@ import org.eyeseetea.malariacare.data.database.datasources.AuthenticationLocalDa
 import org.eyeseetea.malariacare.data.database.datasources.UserAccountDataSource;
 import org.eyeseetea.malariacare.data.database.utils.Session;
 import org.eyeseetea.malariacare.data.remote.AuthenticationDhisSDKDataSource;
+import org.eyeseetea.malariacare.data.remote.ForgotPasswordDataSource;
 import org.eyeseetea.malariacare.domain.boundary.IAuthenticationManager;
 import org.eyeseetea.malariacare.domain.boundary.repositories.IUserRepository;
 import org.eyeseetea.malariacare.domain.entity.Credentials;
@@ -18,12 +19,14 @@ import org.eyeseetea.malariacare.domain.exception.ConfigJsonIOException;
 public class AuthenticationManager implements IAuthenticationManager {
     IAuthenticationDataSource userAccountLocalDataSource;
     IAuthenticationDataSource userAccountRemoteDataSource;
+    ForgotPasswordDataSource mForgotPasswordDataSource;
     IUserRepository mUserRepository;
 
     public AuthenticationManager(Context context) {
 
         userAccountLocalDataSource = new AuthenticationLocalDataSource(context);
         userAccountRemoteDataSource = new AuthenticationDhisSDKDataSource(context);
+        mForgotPasswordDataSource = new ForgotPasswordDataSource(context);
         mUserRepository = new UserAccountDataSource();
     }
 
@@ -62,7 +65,7 @@ public class AuthenticationManager implements IAuthenticationManager {
     @Override
     public void forgotPassword(String username,
             final Callback<ForgotPasswordMessage> callback) {
-        userAccountRemoteDataSource.forgotPassword(username,
+        mForgotPasswordDataSource.forgotPassword(username,
                 new IDataSourceCallback<ForgotPasswordMessage>() {
                     @Override
                     public void onSuccess(ForgotPasswordMessage result) {
