+++ app/src/main/java/org/eyeseetea/malariacare/data/remote/AuthenticationDhisSDKDataSource.java
@@ -6,10 +6,7 @@ import android.net.NetworkInfo;
 
 import org.eyeseetea.malariacare.data.IAuthenticationDataSource;
 import org.eyeseetea.malariacare.data.IDataSourceCallback;
-import org.eyeseetea.malariacare.data.remote.strategies.AuthenticationDhisSDKDataSourceStrategy;
-import org.eyeseetea.malariacare.data.remote.strategies.IAuthenticationDhisSDKDataSourceStrategy;
 import org.eyeseetea.malariacare.domain.entity.Credentials;
-import org.eyeseetea.malariacare.domain.entity.ForgotPasswordMessage;
 import org.eyeseetea.malariacare.domain.entity.UserAccount;
 import org.eyeseetea.malariacare.domain.exception.InvalidCredentialsException;
 import org.eyeseetea.malariacare.domain.exception.NetworkException;
@@ -27,11 +24,9 @@ import rx.schedulers.Schedulers;
 
 public class AuthenticationDhisSDKDataSource implements IAuthenticationDataSource {
     private Context mContext;
-    private IAuthenticationDhisSDKDataSourceStrategy mAuthenticationDhisSDKDataSourceStrategy;
 
     public AuthenticationDhisSDKDataSource(Context context) {
         mContext = context;
-        mAuthenticationDhisSDKDataSourceStrategy = new AuthenticationDhisSDKDataSourceStrategy();
     }
 
     @Override
@@ -56,14 +51,6 @@ public class AuthenticationDhisSDKDataSource implements IAuthenticationDataSourc
             callback.onSuccess(null);
         }
     }
-
-    @Override
-    public void forgotPassword(String username,
-            IDataSourceCallback<ForgotPasswordMessage> callback) {
-        mAuthenticationDhisSDKDataSourceStrategy.forgotPassword(username, isNetworkAvailable(),
-                callback);
-    }
-
 
     @Override
     public void login(final Credentials credentials,
@@ -95,7 +82,8 @@ public class AuthenticationDhisSDKDataSource implements IAuthenticationDataSourc
                         @Override
                         public void call(
                                 org.hisp.dhis.client.sdk.models.user.UserAccount dhisUserAccount) {
-                            UserAccount userAccount = new UserAccount(credentials.getUsername(), dhisUserAccount.getUId(),
+                            UserAccount userAccount = new UserAccount(credentials.getUsername(),
+                                    dhisUserAccount.getUId(),
                                     credentials.isDemoCredentials());
                             callback.onSuccess(userAccount);
                         }
