+++ app/src/main/java/org/eyeseetea/malariacare/data/remote/IForgotPasswordDataSource.java
@@ -1,0 +1,8 @@
+package org.eyeseetea.malariacare.data.remote;
+
+import org.eyeseetea.malariacare.data.IDataSourceCallback;
+import org.eyeseetea.malariacare.domain.entity.ForgotPasswordMessage;
+
+public interface IForgotPasswordDataSource {
+    void forgotPassword(String username, IDataSourceCallback<ForgotPasswordMessage> callback);
+}
